Move Run.transact tests to core/test_run.py
